feat(chat,my-shop): 인증 페이지에 서버 프리페칭 및 서버 인증 가드 적용
